// src/store/actions/types.ts
export enum CLIENTS_ACTIONS {
  GET_CLIENTS = 'GET_CLIENTS',
  GET_CLIENTS_SUCCESS = 'GET_CLIENTS_SUCCESS',
  SAVE_CLIENT = 'SAVE_CLIENT',
  SAVE_CLIENT_SUCCESS = 'SAVE_CLIENT_SUCCESS',
  DELETE_CLIENT = 'DELETE_CLIENT',
  DELETE_CLIENT_SUCCESS = 'DELETE_CLIENT_SUCCESS',
  UPDATE_CLIENT = 'UPDATE_CLIENT',
  UPDATE_CLIENT_SUCCESS = 'UPDATE_CLIENT_SUCCESS',
}